inputkit-webhook-sales/requirements.txt
google-cloud-bigquery==3.19.0
functions-framework==3.5.0